requirements.txt
pillow
numpy
funPIL
PySimpleGUI
playsound==1.2.2